HELPS-123: RG - splitting test groups into its own files
