+++ ldk/javascript/examples/self-test-loop/src/config/user-group.ts
@@ -1,0 +1,8 @@
+import TestGroup from '../testingFixtures/testGroup';
+import { LoopTest } from '../testingFixtures/loopTest';
+
+import * as userTests from '../tests/user';
+
+export const userTestGroup = (): TestGroup => new TestGroup('User Aptitude', [
+    new LoopTest('User Aptitude - JWT', userTests.userJWTTest, 10000, 'No action required'),
+  ]);
